Move clock related stuff to its own store
Was previously in schedules.ts
